feat: add `release upload` command
