+++ commands/release/release.go
@@ -4,6 +4,7 @@ import (
 	"github.com/profclems/glab/commands/cmdutils"
 	releaseCreateCmd "github.com/profclems/glab/commands/release/create"
 	releaseListCmd "github.com/profclems/glab/commands/release/list"
+	releaseUploadCmd "github.com/profclems/glab/commands/release/upload"
 
 	"github.com/spf13/cobra"
 )
@@ -19,6 +20,7 @@ func NewCmdRelease(f *cmdutils.Factory) *cobra.Command {
 
 	releaseCmd.AddCommand(releaseListCmd.NewCmdReleaseList(f))
 	releaseCmd.AddCommand(releaseCreateCmd.NewCmdCreate(f, nil))
+	releaseCmd.AddCommand(releaseUploadCmd.NewCmdUpload(f, nil))
 
 	return releaseCmd
 }
